requirements.txt
flask
flask-mail
flask-wtf
wtforms
email_validator
python-dotenv
requests
